requirements.txt
# Django==4.2.7
# psycopg2-binary==2.9.7
# python-dotenv==1.0.0
# Pillow==10.0.1
# django-tailwind==3.6.0
# razorpay==1.3.0
# openpyxl==3.1.2
# pandas==2.1.3
# requests==2.31.0
# gunicorn==21.2.0
# whitenoise==6.6.0
# django-allauth==0.57.0
# langchain==0.0.350
# openai==1.3.7

Django==4.2.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
openai==1.3.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-allauth==0.57.0
